fix(FJS-4): renamed file in theme directory
